Adding the top-level build system based on autoconf and automake. Possibly including a CMake based build system would also be useful. Here is the REAME file:
--------------------------------------------------------------------
               ____         ____       ____        _     
              |  _ \       / ___|     |___ \      | |    
              | |_) |     | |           __) |     | |    
              |  __/      | |___       / __/      | |___ 
              |_|          \____|     |_____|     |_____|

              A Parallel & Cloud Computing Library (PC2L)
               <http://www.pc2lab.cec.miamioh.edu/pc2l>

---------------------------------------------------------------------
                       ___             _           
                      | _ \___ __ _ __| |_ __  ___ 
                      |   / -_) _` / _` | '  \/ -_)
                      |_|_\___\__,_\__,_|_|_|_\___|


LICENSE

A Parallel & Cloud Computing Library (PC2L) is free software: you can
redistribute it and/or modify it under the terms of the GNU General
Public License (GPL) as published by the Free Software Foundation,
either version 3 (GPL v3), or (at your option) a later version.  You
should have received a copy of the GNU General Public License along
with MUSE.  If not, you may download copies of GPL V3 from
<http://www.gnu.org/licenses/>.


DISCLAIMER

PC2L  is distributed in the hope that it will  be useful,
but   WITHOUT  ANY  WARRANTY;  without  even  the IMPLIED
WARRANTY of  MERCHANTABILITY  or FITNESS FOR A PARTICULAR
PURPOSE.

Miami University and  the PC2Lab development team make no
representations  or  warranties  about the suitability of
the software,  either  express  or implied, including but
not limited to the implied warranties of merchantability,
fitness  for a  particular  purpose, or non-infringement.
Miami  University and  its affiliates shall not be liable
for any damages  suffered by the  licensee as a result of
using, modifying,  or distributing  this software  or its
derivatives.

By using or  copying  this  Software,  Licensee  agree to
abide  by the intellectual  property laws,  and all other
applicable  laws of  the U.S.,  and the terms of the  GNU
General  Public  License  (version 3).


COMPLING

Downloading and Compiling MUSE:

$ svn co --username raodm https://pc2lab.csi.miamioh.edu/svn/pc2lab/pc2l pc2l
$ autoreconf -i
$ ./configure
$ make